--- 755200/29020 755200 游城棣.cpp
@@ -1,11 +1,0 @@
-#include <iostream>
-using namespace std;
-
-int main() {
-  char name[20];
-  cout << "Please your name:\n";
-  cin >> name;
-  cout << "Hello，";
-  cout << name;
-  cout << "\n";
-}
