+++ 755201/29021 755201寧家楨.cpp
@@ -1,0 +1,13 @@
+#include <iostream>
+using namespace std;
+
+int main() {
+    char name[20];
+
+    cout << "½Ð¿é¤J¦W¦r:\n ";
+    cin >> name;
+
+    cout << "Hello¡A";
+    cout << name;
+    cout << "\n";
+}
