+++ 755201/homework form 755201 寧家楨.md
@@ -1,0 +1,8 @@
+關於課程
+為什麼你想來這個課程？你希望這個課程可以讓你達到什麼狀態，或是成為怎樣的人？
+因為有朋友邀我來一起參加這個增廣課，也順便增進自子撰寫程式的能力。
+希望可以撰寫一定難度的程式碼，將需要繁雜程序的事情用程式幫我完成。
+
+關於未來
+你對什麼有興趣？你希望大學可以讀什麼科系？為什麼？
+我對生物有興趣。希望可以讀生物系。因為我喜歡研究動物及古生物。
